Change record sorting to be case insensitve
If one has not set the locale of one's system, sorting of the zone list
takes place using POSIX sorting which puts all of the records starting
with UPPERCASE letters first.  This can be annoying in an environment
where the sysadmins were not systematic about use of case.  The above
patch just forces the sort function to always compare uppercase versions
of the RRs.

Signed-off-by: Erik Rossen <rossen@rossen.ch>
Signed-off-by: Jeff Layton <jlayton@poochiereds.net>